requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-RESTful==0.3.10
Werkzeug==2.3.7

# Database adapters
PyMySQL==1.1.0
psycopg2-binary==2.9.7
jaydebeapi==1.2.3

# Web forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Template and static file handling
Jinja2==3.1.2
MarkupSafe==2.1.3

# HTTP and API support
requests==2.31.0
Flask-CORS==4.0.0

# Basic data processing (for assessment results)
pandas==2.0.3

# Caching
Flask-Caching==2.1.0

# Configuration and environment
python-dotenv==1.0.0
click==8.1.7

# Date and time handling
python-dateutil==2.8.2

# JSON and data serialization
jsonschema==4.19.0

# Logging
structlog==23.1.0

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# Production server
gunicorn==21.2.0

# Math and statistics (for scoring algorithms)
numpy==1.24.4